trying derived types for r/u/v vars!
Initial test to see if I can use 'derived types' for all of the book keeping variables that are declared for rho, u and v.
Old way is to physically copy and rename all variables. With derived types only need to do it once!

It is working, also for an allocatable variable.

This would also be a great solution for netcdf booking variables. Don't have to repeat everytime can just have:

type ncdf_var
      real    :: data_in(GLOBAL_2D_ARRAY,2)
      integer :: grd_type  = 0
      integer :: file_indx = 0
      integer :: irec      = 0
      integer :: it1 = 1, it2 = 2
      real, dimension(2) :: times = [-99,-99]
end type ncdf_var

Then could just use as:

  type (ncdf_var) :: uwnd

  set_var_frc_tile(uwnd, uwnd%data_in, uwnd%grd_type, ... etc )