Fixed error with output frequency.
The output frequency was correct, but the calculations of diagnostics corresponding to that was not correct. Since the calculation needs to be done a step before, and the setting of Hz_old needs to be done a step before that even.

Added some descriptions on how the module works at the top of the module.

Documentation/readme-specifics/README.diagnostics_u_v_value_changes:
- notes on how u changes in timestep and its units.