Selecting specific tracers only is working
Put in functionality to chose which tracers you want to do diagnostics on. This is very important because
a simulation with many tracers would have a gigantic memory footprint and calculation footprint. approx 6xNT 3D arrays...

Confirmed with just salt output that value still the same as salt output with both salt and temp.

This methodology could potentially be used for normal averaging in ROMS, as currently we have t_avg(...,NT),
even if we do not want averages of all tracers. This could substantially reduce the memory footprint for BGC!

diagnostics.F:
- diag_trc logical parameter to move away from cppflags, need to follow through with this still.

t3dmix_GP.F:
- fixed bug of looping through NT when it already called from an NT loop.
- values are bad in t3dmix diagnostic term so need to figure out why.