Better fix of closed boundary.
Previous commit I had hardcoded FlxU & FC to zero, which is incorrect.

Rather the boundary is set to zero now in get_init if the boundary is not open.
This means you don't have to fix the initial file to zero in the boundary offline.

Have left in debuging for now but will strip after this commit. For now can still see FlxU is immediately zero
with OBC_CHECK flag.