Added along bry spatial smoothing and filter
Added along boundary spatial smooting and time filter. The previous commit, the sponge was quite noisy and didn't reach very high or low values from the 0.5 initial condition.

For the sample problem the min/max of the sponge increased, and the sponge was smoother, though does have sharp jumps in places.