particle_join seems to be working
Join tool worked for seeded domain with 9000 particles.

Improved lookup routines to set lower bound to the previously found index for searching
in a sorted list from a sorted list, as next value guaranteed to be greater.

Fixed bug in reading p_in before having established the size of p_in with nparts.